==> notebooks/01_load_and_validate_raw_data.ipynb <==
import pandas as pd
# hay instalar el modulo openpyxl para que pueda leer el archivo xlsx


#load xlsx file
ruta = r'D:\0_Respaldo\0_Proyectos_2024\ML proyects\Prediccion tasa de interes\Prediccion_tasa_interes\data\Serie_historica_ipvnbr.xlsx'
data = pd.read_excel(ruta, 
                     sheet_name='Serie_IPVNBR',
                     header=2,
                     skiprows=2,
                     skipfooter=8)

df = data.copy()

df.head(3)

df.columns

### Data Validation

# Reviso que la fecha este correcta en el perido de tiempo de 2006 a la actualidad.
df['Unnamed: 0'].describe()

Se puede observar que hay fechas menores al 2006 por tal motivo ajustare la fecha.

df = df[df['Unnamed: 0'] >= '2006-01-01']

df1 = df[['Unnamed: 0','Bogotá.1','Alrededores de Bogotá4,5','Medellín.1','Cali.1']]

df1.head(5)

df1 = df1.rename(columns={'Unnamed: 0':'Fecha',
                    'Bogotá.1':'Bogotá',
                    'Alrededores de Bogotá4,5':'Alrededores de Bogotá',
                    'Medellín.1':'Medellín',
                    'Cali.1':'Cali'
                    })

df1.info()

#Valores duplicados
df1.duplicated().sum()

df1.to_parquet(r'D:\0_Respaldo\0_Proyectos_2024\ML proyects\Prediccion tasa de interes\Prediccion_tasa_interes\data\Transform\validate_data.parquet')
