==> notebooks/02_trasnform_raw_data_into_ts_data.ipynb <==
import pandas as pd

data = pd.read_parquet(r'D:\0_Respaldo\0_Proyectos_2024\ML_proyects\Prediccion tasa de interes\Prediccion_tasa_interes\data\Transform\validate_data.parquet')

df = data.copy()

df.head(5)

### Bucle 

c_list = ['Bogotá', 'Alrededores de Bogotá', 'Medellín', 'Cali']

#Bucle
for i in c_list:
    df_place = df[['Fecha', i]]
    path = r'D:\0_Respaldo\0_Proyectos_2024\ML_proyects\Prediccion tasa de interes\Prediccion_tasa_interes\data\Transform\ts_'
    df_place.to_parquet(f'{path}{i}.parquet')

###### Month Range creation example.

# for month in range(1, 13):
#     month_name = pd.Timestamp(f'2023-{month:02d}-01').strftime('%B').lower()  # Nombre del mes en minúscula
#     df1[f'mes {month_name}'] = df['Fecha'].apply(lambda x: 1 if x.month == month else 0)
    
    
# # Crear semanas
# for week in range(1, 53):
#     df1[f'semana {week}'] = df['Fecha'].apply(lambda x: 1 if x.isocalendar().week == week else 0)
    
# # Crear  Trimestres

# for quarter in range(1, 5):
#     df[f'trimestre {quarter}'] = df['Fecha'].apply(lambda x: 1 if x.quarter == quarter else 0)
