==> notebooks/03_transform_ts_data_into_features_and_targets.ipynb <==
import pandas as pd

from tqdm import tqdm  #### Loading work bar

import numpy as np

# Funcion para cortar indices y cerar los dataframes
def get_cutoff_indices(
    data: pd.DataFrame,
    n_features: int,
    step_size: int
    ) -> list:

        stop_position = len(data) - 1
        
        # Start the first sub-sequence at index position 0
        subseq_first_idx = 0
        subseq_mid_idx = n_features
        subseq_last_idx = n_features + 1
        indices = []
        
        while subseq_last_idx <= stop_position:
            indices.append((subseq_first_idx, subseq_mid_idx, subseq_last_idx))
            
            subseq_first_idx += step_size
            subseq_mid_idx += step_size
            subseq_last_idx += step_size

        return indices

### Bucle

ciudades = ['Bogotá', 'Alrededores de Bogotá', 'Medellín', 'Cali']


for i in ciudades:
    ruta = r'D:\0_Respaldo\0_Proyectos_2024\ML_proyects\Prediccion tasa de interes\Prediccion_tasa_interes\data\Transform\ts_'
    df_c = pd.read_parquet(f'{ruta}{i}.parquet')
    df_c.reset_index(drop=True, inplace=True)  # Asegúrate de que los índices se restablecen

    n_features = 12
    step_size = 1

    indices = get_cutoff_indices(df_c, n_features, step_size)
    
    n_examples = len(indices)
    x = np.ndarray(shape=(n_examples, n_features), dtype=np.float32)
    y = np.ndarray(shape=(n_examples), dtype=np.float32)
    
    months = []

    for u, idx in enumerate(indices):
        x[u, :] = df_c.iloc[idx[0]:idx[1]][i].values
        y[u] = df_c.iloc[idx[1]:idx[2]][i].values  # Asegúrate de seleccionar la columna correcta
        months.append(df_c.iloc[idx[1]]['Fecha'])
        
    feature = pd.DataFrame(x, columns=[f'month_{j+1}' for j in range(n_features)])
    
    target = pd.DataFrame(y, columns=['Target'])
    
    # Crear el DataFrame con los datos concatenados.
    df_ready = pd.concat([feature, target], axis=1)
    
    # Exportar los datos al archivo train
    ruta = r'D:\0_Respaldo\0_Proyectos_2024\ML_proyects\Prediccion tasa de interes\Prediccion_tasa_interes\data\Transform\train\train_'
    df_ready.to_parquet(f'{ruta}{i}.parquet')
    print(f'Locación: {i}')
    print(df_ready.head(3))
