==> notebooks/04_ligghtGBM_ML.ipynb <==
import pandas as pd
import lightgbm as lgb

from sklearn.model_selection import train_test_split
from sklearn.preprocessing import StandardScaler
from sklearn.metrics import mean_squared_error, mean_absolute_error

### ML Bogota

data = pd.read_parquet(r'D:\0_Respaldo\0_Proyectos_2024\ML_proyects\Prediccion tasa de interes\Prediccion_tasa_interes\data\Transform\train\train_Bogotá.parquet')
data.head(3)

df =  data.copy()

# Preparacion de los datos
x = df.drop(['Target'], axis=1) #Features
y = df['Target'] #target

x_train, x_test, y_train, y_test = train_test_split(x, y, test_size = 0.2, random_state = 28)

model = lgb.LGBMRegressor()

model.fit(
    x_train, y_train,
    eval_set=[(x_test, y_test)],
    eval_metric='rmse',    
)

#Model prediction

y_pred = model.predict(x_test)
predictions = model.predict(x_test)

rmse = mean_squared_error(y_test, y_pred, squared=False)
test_mae = mean_absolute_error(y_test, predictions)

print(f'{test_mae=:.4f}')
print(f'RMSE: {rmse=:.4f}')

import joblib as jb

jb.dump(model, r'D:\0_Respaldo\0_Proyectos_2024\ML_proyects\Prediccion tasa de interes\Prediccion_tasa_interes\models\ml_bogota.pkl' )

### ML Alrededores_bogota

data = pd.read_parquet(r'D:\0_Respaldo\0_Proyectos_2024\ML_proyects\Prediccion tasa de interes\Prediccion_tasa_interes\data\Transform\train\train_Alrededores de Bogotá.parquet')
data.head(3)

df =  data.copy()
# Preparacion de los datos
x = df.drop(['Target'], axis=1) #Features
y = df['Target'] #target

x_train, x_test, y_train, y_test = train_test_split(x, y, test_size = 0.2, random_state = 28)

model = lgb.LGBMRegressor()

model.fit(
    x_train, y_train,
    eval_set=[(x_test, y_test)],
    eval_metric='rmse',    
)

#Model prediction

y_pred = model.predict(x_test)
predictions = model.predict(x_test)

rmse = mean_squared_error(y_test, y_pred, squared=False)
test_mae = mean_absolute_error(y_test, predictions)

print(f'{test_mae=:.4f}')
print(f'RMSE: {rmse=:.4f}')

jb.dump(model, r'D:\0_Respaldo\0_Proyectos_2024\ML_proyects\Prediccion tasa de interes\Prediccion_tasa_interes\models\ml_alrededores_bogota.pkl' )

### ML Medellin

data = pd.read_parquet(r'D:\0_Respaldo\0_Proyectos_2024\ML_proyects\Prediccion tasa de interes\Prediccion_tasa_interes\data\Transform\train\train_Medellín.parquet')
data.head(3)

df =  data.copy()
# Preparacion de los datos
x = df.drop(['Target'], axis=1) #Features
y = df['Target'] #target

x_train, x_test, y_train, y_test = train_test_split(x, y, test_size = 0.2, random_state = 28)

model = lgb.LGBMRegressor()

model.fit(
    x_train, y_train,
    eval_set=[(x_test, y_test)],
    eval_metric='rmse',    
)

#Model prediction

y_pred = model.predict(x_test)
predictions = model.predict(x_test)

rmse = mean_squared_error(y_test, y_pred, squared=False)
test_mae = mean_absolute_error(y_test, predictions)

print(f'{test_mae=:.4f}')
print(f'RMSE: {rmse=:.4f}')

jb.dump(model, r'D:\0_Respaldo\0_Proyectos_2024\ML_proyects\Prediccion tasa de interes\Prediccion_tasa_interes\models\ml_medellin.pkl' )

### Cali

data = pd.read_parquet(r'D:\0_Respaldo\0_Proyectos_2024\ML_proyects\Prediccion tasa de interes\Prediccion_tasa_interes\data\Transform\train\train_Cali.parquet')
data.head(3)

df =  data.copy()
# Preparacion de los datos
x = df.drop(['Target'], axis=1) #Features
y = df['Target'] #target

x_train, x_test, y_train, y_test = train_test_split(x, y, test_size = 0.2, random_state = 28)

model = lgb.LGBMRegressor()

model.fit(
    x_train, y_train,
    eval_set=[(x_test, y_test)],
    eval_metric='rmse',    
)

#Model prediction

y_pred = model.predict(x_test)
predictions = model.predict(x_test)

rmse = mean_squared_error(y_test, y_pred, squared=False)
test_mae = mean_absolute_error(y_test, predictions)

print(f'{test_mae=:.4f}')
print(f'RMSE: {rmse=:.4f}')

jb.dump(model, r'D:\0_Respaldo\0_Proyectos_2024\ML_proyects\Prediccion tasa de interes\Prediccion_tasa_interes\models\ml_cali.pkl' )
