==> notebooks/06_backfill_feature_store.ipynb <==
HOPSWORKS_PROJECT_NAME = 'tasa_de_interes_colombia'

import os
from dotenv import load_dotenv

load_dotenv(r'D:\0_Respaldo\0_Proyectos_2024\ML_proyects\Prediccion tasa de interes\Prediccion_tasa_interes\src\.env')

HOPSWORKS_API_KEY = os.environ['HOPSWORKS_API_KEY']
